Address more comments from PR
