+++ state/protocol/badger/mutator.go
@@ -4,9 +4,8 @@ package badger
 
 import (
 	"fmt"
-	"github.com/onflow/flow-go/state/protocol"
-
 	"github.com/dgraph-io/badger/v2"
+	"github.com/onflow/flow-go/module"
 
 	"github.com/onflow/flow-go/model/flow"
 	"github.com/onflow/flow-go/module/trace"
@@ -17,7 +16,7 @@ import (
 
 type Mutator struct {
 	state     *State
-	validator protocol.ReceiptValidator
+	validator module.ReceiptValidator
 }
 
 func (m *Mutator) Bootstrap(root *flow.Block, result *flow.ExecutionResult, seal *flow.Seal) error {
